# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Planner API documentation",
        "description": "API Documentation for Travel Planner",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/itinerary": {
            "get": {
                "tags": [
                    "Itinerary"
                ],
                "summary": "Get all itineraries for logged in user",
                "description": "Get all itineraries for logged in user",
                "operationId": "get_itinerary",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Itinerary"
                ],
                "summary": "Create a itinerary for a destination",
                "description": "Create itinerary",
                "operationId": "create_itinerary",
                "parameters": [
                    {
                        "name": "destination_id",
                        "in": "query",
                        "description": "required | exists:destinations,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "required | date_format:Y-m-d | after:today",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "properties": {
                                "schedule": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "activity": {
                                                "type": "string"
                                            },
                                            "cost": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itinerary/update": {
            "post": {
                "tags": [
                    "Itinerary"
                ],
                "summary": "Update itinerary",
                "description": "Update itinerary",
                "operationId": "update_itinerary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "required | exists:itineraries,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "date_format:Y-m-d | after:today",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "properties": {
                                "schedule": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "activity": {
                                                "type": "string"
                                            },
                                            "cost": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itinerary/delete": {
            "post": {
                "tags": [
                    "Itinerary"
                ],
                "summary": "Delete itinerary",
                "description": "Delete itinerary",
                "operationId": "delete_itinerary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "required | exists:itineraries,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/batch/itinerary": {
            "post": {
                "tags": [
                    "Itinerary"
                ],
                "summary": "Create multiple itineraries for a destination",
                "description": "Create multiple itineraries for a destination",
                "operationId": "batch_create_itinerary",
                "parameters": [
                    {
                        "name": "destination_id",
                        "in": "query",
                        "description": "required | exists:destinations,id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "properties": {
                                "itineraries": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "date": {
                                                "type": "string"
                                            },
                                            "schedule": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "activity": {
                                                            "type": "string"
                                                        },
                                                        "cost": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Error"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}